chore: user's game status button has been added
